# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/settings/{category}": {
            "get": {
                "description": "根据分类名称获取该分类下所有设置项的键值、类型与描述，未入库项返回默认值",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统设置接口"
                ],
                "summary": "[管理] 获取分类设置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "设置分类(如 general/qa/repo/security)",
                        "name": "category",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/settings.CategorySettingsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "无效的设置分类",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "批量更新指定分类下的设置项键值，校验分类合法性、Key 归属与值类型后逐项写入",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统设置接口"
                ],
                "summary": "[管理] 更新分类设置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Access Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "设置分类(如 general/qa/repo/security)",
                        "name": "category",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新分类设置请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/settings.UpdateCategorySettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "设置更新成功",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误或设置项不属于该分类",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/common.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "common.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "context": {
                    "type": "string"
                },
                "data": {},
                "error_message": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "output": {
                    "type": "string"
                },
                "overhead": {
                    "type": "integer"
                }
            }
        },
...
        "settings.CategorySettingsResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "分类名称(如general/repo/memory)",
                    "type": "string"
                },
                "items": {
                    "description": "该分类下的设置项列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/settings.SettingItemResponse"
                    }
                }
            }
        },
        "settings.SettingItemResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "设置项描述说明",
                    "type": "string"
                },
                "key": {
                    "description": "设置项键名",
                    "type": "string"
                },
                "type": {
                    "description": "值类型(text/number/boolean等)",
                    "type": "string"
                },
                "value": {
                    "description": "设置项当前值",
                    "type": "string"
                }
            }
        },
        "settings.UpdateCategorySettingsRequest": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "待更新的键值对列表",
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "key": {
                                "description": "设置项键名",
                                "type": "string"
                            },
                            "value": {
                                "description": "新的设置值",
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    }
}